requirements.txt
# GCP
google-cloud-bigquery==3.17.2
google-cloud-storage==2.16.0
google-cloud-secret-manager==2.20.0
google-cloud-logging==3.10.0

# Spark / Ingestion
pyspark==3.5.1
pandas==2.2.1
numpy==1.26.4
pyarrow==15.0.2
yfinance==0.2.38
fredapi==0.5.1
requests==2.31.0

# AI / RAG
langchain==0.3.20
langchain-anthropic==0.3.10
langchain-community==0.3.19
langchain-google-community==2.0.7
langgraph==0.2.73
langsmith==0.2.10
google-cloud-aiplatform==1.57.0
rank-bm25==0.2.2

# API
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.4
python-multipart==0.0.9
httpx==0.27.0
user-agents==2.2.0

# Utilities
python-dotenv==1.0.1
tenacity==8.3.0
